requirements.txt
adblockparser==0.7
asgiref==3.3.4
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
constantly==15.1.0
Django==3.2
djangorestframework==3.12.4
funcparserlib==0.3.6
greenlet==1.0.0
gunicorn==20.1.0
h2==3.2.0
hpack==3.0.0
html5lib==1.1
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
Pillow==8.2.0
priority==1.3.0
psutil==5.8.0
psycopg2==2.8.6
python-dotenv==0.17.0
pytz==2021.1
qt5reactor==0.6.3
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
splash==3.5
SQLAlchemy==1.4.11
sqlparse==0.4.1
Twisted==21.2.0
urllib3==1.26.4
webencodings==0.5.1
whitenoise==5.2.0
xvfbwrapper==0.2.9
zope.interface==5.4.0
